requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
dateutils==0.6.6
decorator==4.4.0
Django==2.1.7
django-filter==2.1.0
django-rest-framework-braces==0.3.4
django-storages==1.7.1
djangorestframework==3.9.2
fcsparser==0.2.0
FlowCytometryTools==0.5.0
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
idna==2.8
Markdown==3.0.1
numpy==1.16.2
pandas==0.24.2
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
